requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-SocketIO==5.4.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
pygame==2.6.1
python-engineio==4.10.1
python-socketio==5.11.4
simple-websocket==1.1.0
Werkzeug==3.1.3
wsproto==1.2.0
